sheep: remove vdi_deleted from vdi state
When we delete a vdi, its inode will remain in the backend store because we need
it to indicate if this bit in vdi_bitmap is used or not because we use bitmap
for vdi allocation as following:

             bit name
created       1   Yes
deleted       1   No
Not-hashed    0   N/A

If the bit is once allocated, it will never get cleared because of handling of
hash collision problem.

There is no need to keep track of deletion state to acclerate the 'dog vdi list'
command because the new vdi might reuse the bit if it hashes to.

This is one step forward to remove vdi state from sheepdog.
   
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>