sheep: remove vdi state sync-up
This is the last step to remoe vdi state sync-up stuff. It was my recation by
following conversation:

http://lists.wpkg.org/pipermail/sheepdog/2015-May/067736.html

It is, unfortunately, the start of sheepdog-ng, a sheepdog fork, because two
conflicted ideas can't coexist for now.

sheepdog-ng features a simpler sheepdog, which should focus on storage only and
make a good tradeoff between tuple {stability, scalability, performance,
features}.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>