remove block_size_shift completely
This idea to make object size variable is great, but the implementation is ugly
and mess the code up.

This patch remove related code and just keep block_size_shift field to keep
backward compatibility. So we only support 4MB object size as before. We could
write a new implementation if vairable object size is really called for.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>