remove recovery throttling
The recovery throttling is targeted to accelerate the user requests but in
the real world, this patch will make things worse.

The nature of sheepdog data placement is to distribute evently objects into
different nodes as much as possible. With this assumption, every node will
have all the partial data of every vdi. So if we can't recvery fast, all the
user request will be blocked by the data that is waiting to be recovered.

This is a embrassing dilemma. User requests and recovery requests are
mutually affected and we need think twice before handling this kind of problem,
otherwise, the effect will be negative.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>