sheep: remove tree store
This idea of tree store is to improve performance of large objects, assuming
that many files in a single directory would be a bottleneck. This assumption
is right about old file systme like ext2, but is completely wrong when we talk
about contemporary fs like ext4 or xfs. Take ext4 for example, it uses Htree (a
btree variant) to manage directory entries, meaning it already presents us a
true tree-like structure for file look-ups.

What tree store does is just add tree structure on a tree structure, it won't
help performance at all, but just add complexity of code for sheep.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>