sheep: fix deadlock while in vdi deletion and recovery
how it happens:
1. start delete one of vdi
2. trigger a node event and start recovery
3. deadlock will happen soon.

The reason is:

deletion is executed in block thread:
for_each_object: sd_unrefcnt_object() -> wait on oid that is not recovered

recovery is executed in rw thread:
preprare_object_list() -> wait_get_vdi_bitmap_done() -> wait get_vdi_bitmap()
                                            
get_bitmap is queued in block thread and wait for deletion finishes -> deadlock!

The fix is rather simple, just queue get_bitmap in the recovery thread as it
should be. This is a bug of 3 years old.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>