shared lib: add auto-reconnect support for the client
Add SO_RCVTIMEO and SO_SNDTIMEO option for the socket for network
read and write. The function net_{read,write} may be blocked for
the set TIMEOUT at most(however, for some corner case, it may block
as long as 2 or 3 times of that), or it will return error after
timeout, which implies the remote sheep node may be died. In this
case the client will try to connect to a new sheep node until a
new sheep node is available. After that all the requests in the
inflight queue, in which requests have be sent but the response
haven't arrived, are re-sent to the new connected sheep node.

No matter whether the client is idle or busy for doing I/O, the
client can handle it gracefully so long as the cluster works normally.

Signed-off-by: Yu Yang <yuyang@cmss.chinamobile.com>
[ Reordered some lines - Yuan ]
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>