shared lib: correct direct dependencies in .so need section
If we don't set LIBS, we will inherit from upper layer. So if we are configured
with zookeeper, libsheepdog.so will depend on many ununsed dependencies like
following:

# ldd -u -r /usr/lib64/libsheepdog.so
Unused direct dependencies:
        
        /usr/lib/libzookeeper_mt.so.2
        /lib64/libm.so.6
        /lib64/librt.so.1

Note, --as-needed doesn't work on Centos 6 and I don't know why. So the simple
fix is the exclude unnecessary libs out of lib/* and put what is needed in the
LIBS directly

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>