sheep: fix false warn/error logging messages
Because of thin-provision of ledger object, it might not exist before first
acess, so read request could return SD_RES_NO_OBJ and this is a valid return
code.

The caller of sd_read_object should decide the error message.

The tricky part is that we should return ERROR to the client if the request is
not local.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>