tool/zk_control: fix do_kill
The ZooKeeper FAQ entry titled "Is there an easy way to expire a session for 
testing?" explains that when one client closes a session, all other sessions
using that same session ID will be expired.

So we should create a new session for the target path and then close it.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>