zookeeper: exit when receive EPOLLOUT
It is impossible for us to handle EPOLLOUT since we pass only EPOLLIN to epoll,
meaning that we only monitor EPOLLIN event. But the broken kernel forces us to
handle EPOLLOUT, which means the kernel is insane we'd better do suiside.

This might sounds impossible to you and you might take this a unnecessary
defensive patch, but we do meet this problem in our production system on centos
6.5, weird.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>