sheep: don't start recovery for gateway-only node event
gateway-only node event doesn't change the vnode hash ring at all, so we don't
need to start recovery when this happens. The tricky part is that, when the node
is in recovery, we should not skip the reocvery because the reocvery code run
with the old epoch. We simply start the next recovery to replace the current one
and in this way we make sure epoch is up-to-date.

Note, the next recovery will replace the current one immediately and no wait at
all and all the objects recovered by the last recovery will stay at the working
directory, so the recovery replacement is alsmot zero side effect.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>