sheep: don't interpret random data stream
If someone else writes data to the port sheep listens, sheep will try to process
this data and lead to the effect similar to DDOS, crash the sheep or anything
you can imagine.

This patch only allows sheep to interpret the data in the format of sheepdog
protocol and close the connection if not.

Our check_hrd is much stricter than purely comparing the magic number, meaning a
very low chance of random data would pass it.

And even the faked request pass the check_hrd, it shold not cause any serious
problem such as gateway repeats a should-fail request.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>