sheep/vdi: make deletion asynchronous
We need to put deletion into a worker thread otherwise we'll block
.process_work. So return of the deletion request only implies the success
of the acception of 'vdi delete' request. The real deletion is delayed and
could fail. In the failure case, users are expected to resend the deletion
request.

This attack the problem that deletion of vdi takes too long and block all the
other cluster wide operations (e.g, vdi create)

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>