sheep/recovery: stop rollback recover on err SD_RES_NO_SPACE
There is no need to rollback when we get SD_RES_NO_SPACE, just try to recover
the next oid. This is just a partial solution trying to accelerate the recovery
process while we get SD_RES_NO_SPACE. SD_RES_NO_SPACE remains unhandled in the
recovery, need to be fixed in the future.

Signed-off-by: niko <tongfw@gmail.com>
[ Add the commit log ]
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>