sheep: introduce node status
Node status affect manual recovery code path only. If SD_CLUSTER_FLAG_MANUAL
not set, the status of all the nodes in the cluser is RUNNING.

Node is online:

1) RUNNING - can server both read and write
2) RECOVER - can server only write

Node is offline:

Neither read nor write could be sent to this node but the remaining nodes are
not in recovery, meaning the data replica on this node will not be recovered
immediately. Instead, we exepct this node to come back. Once the node is back
the recovery of the delta data during the downtime takes place only on this
node as opposed to auto-recovery, that is, node event will cause the whole
cluster to be in recovery.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>