sheep: introduce last_status
If the cluster is shutdowned, the restart of the nodes that tries to join the
cluster should be distiguished from the nodes that rejoin the running cluster.
The first case doesn't need recover while the latter one need. We introduce
last_status to assist us to decide what the cluster state is when we join a node

last_status: WAIT, status: OK --> The cluster is restarting
last_status: OK, status: OK --> Just a node joins a running cluster

With last_status, we can have finer node status control:

A joining node can either be RECOVER or RUNNING and only after recovery, RECOVER
can be made to RUNNING.

If all nodes of the cluster are shutdowned cleanly, there is no recovery after
startup.

For power failure, after all nodes restart, all the nodes will be in recovery
which is the same as aut-recovery to handle power failure at recovery.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>