tests/functional: fix spurious failure caused by c3e5e91
c3e5e91 (sheep: put the request on to sleep queue while doing zk reconnecting, 2016-02-29)

This commit put the request on the sleep queue if cluster is waiting for other
node to join and cause 52 and 88 will hang forever because of bad command.

Just remove the wrong commands fix the problem.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>