sheep: don't touch the epoch time when copied from remote
The command 'dog cluster info' shows epoch history, when local sheep
lacks an epoch, it will pull the epoch file from remote sheep.
After pulled from remote, local sheep will create a local copy of
the epoch file with current local time as timestamp of the epoch.

Actually the timestamp of the epoch should reveal the time of the
epoch event, so we should retain the timestamp of the epoch from remote
rather than update it. Fix it.

Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>