zookeeper: don't loop on ZCONNECTIONLOSS endlessly
If zk cluster is out of service for any reason, sheep will get ZCONNECTIONLOSS
if it wants any service from zookeeper. It doesn't make sense for sheep to loop
endlessly on ZCONNECTIONLOSS, resulting in the hang of the sheep in the main-
thread and further hang the cluster since no one else can kick this hung sheep
out of sheep cluster.

This can be easily reproduced:

1 start 3 zks
2 create a 3 nodes sheep cluster (-c 3) and create a vdi named test
3 dd if=/dev/urandom | $dog vdi write test
3 kill 2 zks and then the zk cluster goes out of service
4 $dog vdi create test1 # Will cause the connected sheep to be hung up

Then you will notice that step #3, the writing of the test is hung up and
any new connections to the hung-up sheep will be refused to connect in the
sheep log.

The fix is rather simple, we just remove ZCONNECTIONLOSS from the while loop
because:

1 Without zk, only cluster-wide operation(vdi create, etc.) fails and we should
  just fail out but normal IOs from client could still be executed if there is
  no membership change.
2 Once zk cluster resume alive, zk client will handle reconnection interlally
  and we can handle session expire (if any) gracefully

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>