md: fix stale corrupted error upon mountpoint loss
As we encountered, when we are on handling ENOENT error under the .stale directory,
and the log appears like:
	ERROR [gway 121713] err_to_sderr(61) /shd/obj10/.stale corrupted
	ERROR [gway 121743] err_to_sderr(61) /shd/obj10/.stale corrupted
	ERROR [gway 121710] err_to_sderr(61) /shd/obj10/.stale corrupted
	...
The ENOENT error is because that the disk is umounted by systemd due to certain
errors found on disk.
But sheep should not report the corruption of a .stale directory.
Actually sheep intends to check the existence of the mountpoint directory:
	if (stat(dir, &s) < 0) {
		sd_err("%s corrupted", dir);
		return md_handle_eio(dir);
	}
But here the "dir" is a .stale directory, and then md_handle_eio() will
check whether the disk exist by comparing string "/shd/obj10/.stale" and
all the disk names like "/shd/obj10", "/shd/obj9", etc.
So it will not be able to find the bad disk and give up.

To fix it, we trim the ".stale" with an extra dirname() call, then
md_handle_eio() will be given an searchable disk name.

Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>