md: add fsid check for sheep store disks to handle mountpoint loss
As we encountered, systemd may umount a disk when it detects certain errors,
then sheep fall back to use the disk(often the system disk) of the mountpoint
directory as a store disk.
As follows:
before
    Id  Size    Used    Avail   Use%    Path
     0  3.6 GB  1.0 GB  2.5 GB   28%    /mnt/sheep1
     1  3.6 GB  296 MB  3.3 GB    8%    /mnt/sheep2 <-- original disk
after
    Id  Size    Used    Avail   Use%    Path
     0  3.6 GB  1.0 GB  2.5 GB   28%    /mnt/sheep1
     1  12 GB   0.0 MB  12 GB     0%    /mnt/sheep2 <-- fall to system disk

Upon this case, sheep fails to tell that the original disk is gone just
based on the directory name.

We add a fsid field to struct disk to check for mountpoint loss.
After the change, the disk sheep fall backed on will be correctly unpluged.

Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>