logger: fix semop return invalid arguemnt when try to start sheep twice
If we try to start a new sheep while having a running sheep,
the new sheep will get a semaphore to protect logarea, later it find
that the 7000 port is taken, so it will free the semaphore.
But the new sheep actually frees the semaphore owned by the running
sheep together, then following log operation all get a EINVAL return.

Because semget() function use a semkey, generated by random() function,
to get a semaphore, and the random() function is known to return
the same random number upon each call, so the new sheep get the same
semaphore as the running sheep.

To fix it, we just use IPC_PRIVATE as semkey, then each sheep
will get a different semaphore on startup.

Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>