dog: add inferface to list the single fully matched vdi quickly
As designed before, it will read all the vdi inodes even if a vdiname
is specified. It's time-consuming, especially when the cluster is in
recovery and inodes are not cached in memory.

1. "dog vdi list" will list all the vdis including snapshots;
2. "dog vdi list requested_name" will list the vdi and all its snapshots;
They are useful, and need to iterate the vdi_inuse.

Add the new option:
dog vdi list vdi_name -s tag -S
could list the single fully matched vdi or snapshot by reading inode
directly.

Signed-off-by: Meng Lingkun <menglingkun@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>