sheep: handle VID overflow correctly
Current sheep cannot handle a case of VID overflow correctly. The main
cause of this problem is invalid usage of fill_vdi_info_range() by
fill_vdi_info() in a case of overflow. This patch fixes this problem.

In addition, this patch lets 0x000000 be an invalid VID. It is for VID
recycling. In some places of sheepdog, parent_vid is used as a
value which indicates that the VDI has a parent (clone or snapshot) or
not (working VDI). So this patch lets the first VID 0x000001 and
prevent this sort of confusion.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Meng Lingkun <menglingkun@cmss.chinamobile.com>
Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>